python_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
supabase==2.18.1
python-multipart==0.0.20
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1